Registration groups + decorative blocks
Registration groups are a mini version of the registration class, designed to let you group similar blocks/items into smaller groups to keep code cleaner

Decorative blocks started, with a system to pass in a model with "texture sets" - a key/value pair to match textures to the definition in the model file without child mods depending on the datagen system to find them.

Rotation dec stuff isn't finished yet!